refactor(api): move assessment and answer serializers
