+++ api/src/evaluation/infrastructure/serializers/jsonapi/answer-serializer.js
@@ -1,6 +1,6 @@
-import { Answer } from '../../../../src/evaluation/domain/models/Answer.js';
+import { Answer } from '../../../domain/models/Answer.js';
 import jsonapiSerializer from 'jsonapi-serializer';
-import { AnswerStatusJsonApiAdapter } from '../../adapters/answer-status-json-api-adapter.js';
+import { AnswerStatusJsonApiAdapter } from '../../../../../lib/infrastructure/adapters/answer-status-json-api-adapter.js';
 
 const { Serializer } = jsonapiSerializer;
