+++ api/src/shared/application/assessments/assessment-controller.js
@@ -6,7 +6,7 @@ import { usecases as devcompUsecases } from '../../../devcomp/domain/usecases/in
 import * as events from '../../../../lib/domain/events/index.js';
 import { logger } from '../../../../lib/infrastructure/logger.js';
 import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
-import * as assessmentSerializer from '../../../../lib/infrastructure/serializers/jsonapi/assessment-serializer.js';
+import * as assessmentSerializer from '../../infrastructure/serializers/jsonapi/assessment-serializer.js';
 import * as activitySerializer from '../../../../lib/infrastructure/serializers/jsonapi/activity-serializer.js';
 import * as challengeSerializer from '../../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
 import * as competenceEvaluationSerializer from '../../../evaluation/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
